requirements.txt
python-dotenv
openai
setuptools